Revert "added ipc"
This reverts commit 1a4733ca7c77d81571418f70b3b7e7b0bac95973.

This commit was a patch https://dwm.suckless.org/patches/ipc/.
Inspecting it a bit better, it is REALLY bad. For one, the IPC works
through a binary protocol, and the serialization is through
__attribute((packed)), which is not a good way to serialize binary data.

Internal state of the program should NOT be shared as an IPC medium, it
should only be used in cases the software works like a "thread" of the
main software. Everything else should be serialized, even if the medium
is a binary.